web-extension/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
google-adk>=1.0.0
litellm>=1.50.0
python-dotenv>=1.0.0
